# gen/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email:   
    name:     
    url: https://josue-lubaki.ca
  description: Springular REST API
  license:
    name: Apache License Version 2.0
  title: Springular REST API
  version: 1.0.0
servers:
- description: Base Url
  url: http://localhost:8080
...
paths:
...
  /api/v1/reservations:
    get:
      operationId: getAllReservationsUsingGET
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/ReservationDTORes'
                type: array
          description: Successfully retrieved all reservations of User
      security:
      - JWT:
        - global
      summary: Get all reservations
      tags:
      - Reservation Controller
      x-accepts: '*/*'
    post:
      operationId: createReservationUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationDTOReq'
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReservationDTORes'
          description: OK
        "201":
          description: Successfully save a reservation
      security:
      - JWT:
        - global
      summary: Save a reservation
      tags:
      - Reservation Controller
      x-contentType: application/json
      x-accepts: '*/*'
  /api/v1/reservations/accept/{id}:
    patch:
      operationId: acceptReservationUsingPATCH
      parameters:
      - description: Reservation ID to accept
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationDTOReq'
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReservationDTORes'
          description: Successfully accepted a reservation
      security:
      - JWT:
        - global
      summary: Accept a reservation
      tags:
      - Reservation Controller
      x-contentType: application/json
      x-accepts: '*/*'
  /api/v1/reservations/{id}:
    delete:
      operationId: deleteReservationUsingDELETE
      parameters:
      - description: Reservation ID to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
        "204":
          description: Successfully deleted a reservation
      security:
      - JWT:
        - global
      summary: Delete a reservation
      tags:
      - Reservation Controller
      x-accepts: application/json
    get:
      operationId: getReservationByIdUsingGET
      parameters:
      - description: Reservation ID to consult
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReservationDTORes'
          description: Successfully retrieved a reservation by id
      security:
      - JWT:
        - global
      summary: Get reservation by id
      tags:
      - Reservation Controller
      x-accepts: '*/*'
    put:
      operationId: updateReservationUsingPUT
      parameters:
      - description: Reservation ID to update
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationDTOReq'
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReservationDTORes'
          description: Successfully updated a reservation
      security:
      - JWT:
        - global
      summary: Update a reservation
      tags:
      - Reservation Controller
      x-contentType: application/json
      x-accepts: '*/*'
  /management/api/v1/reservations:
    get:
      operationId: getAllReservationsUsingGET_1
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/ReservationDTORes'
                type: array
          description: Successfully retrieved all reservations of User
      security:
      - JWT:
        - global
      summary: Get all reservations
      tags:
      - Reservation Management Controller
      x-accepts: '*/*'
    post:
      operationId: createReservationUsingPOST_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationDTOReq'
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReservationDTORes'
          description: OK
        "201":
          description: Successfully save a reservation
      security:
      - JWT:
        - global
      summary: Save a reservation
      tags:
      - Reservation Management Controller
      x-contentType: application/json
      x-accepts: '*/*'
  /management/api/v1/reservations/{id}:
    delete:
      operationId: deleteReservationUsingDELETE_1
      parameters:
      - description: Reservation ID to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
        "204":
          description: Successfully deleted a reservation
      security:
      - JWT:
        - global
      summary: Delete a reservation
      tags:
      - Reservation Management Controller
      x-accepts: application/json
    get:
      operationId: getReservationByIdUsingGET_1
      parameters:
      - description: Reservation ID to consult
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReservationDTORes'
          description: Successfully retrieved a reservation by id
      security:
      - JWT:
        - global
      summary: Get a reservation by id
      tags:
      - Reservation Management Controller
      x-accepts: '*/*'
    put:
      operationId: updateReservationUsingPUT_1
      parameters:
      - description: Reservation ID to update
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationDTOReq'
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReservationDTORes'
          description: Successfully updated a reservation
      security:
      - JWT:
        - global
      summary: Update a reservation
      tags:
      - Reservation Management Controller
      x-contentType: application/json
      x-accepts: '*/*'
...
components:
  schemas:
    AddressDTO:
      example:
        zip: zip
        city: city
        street: street
        id: 0
        state: state
        apartement: apartement
      properties:
        apartement:
          type: string
        city:
          type: string
        id:
          format: int64
          type: integer
        state:
          type: string
        street:
          type: string
        zip:
          type: string
      title: AddressDTO
      type: object
...
    HaircutDTO:
      example:
        estimatedTime: estimatedTime
        price: 0
        imageURL: imageURL
        description: description
        id: id
        title: title
      properties:
        description:
          type: string
        estimatedTime:
          type: string
        id:
          type: string
        imageURL:
          type: string
        price:
          format: int32
          type: integer
        title:
          type: string
      title: HaircutDTO
      type: object
...
    LocationDTO:
      example:
        latitude: 5.962133916683182
        id: 1
        longitude: 5.637376656633329
      properties:
        id:
          format: int64
          type: integer
        latitude:
          format: double
          type: number
        longitude:
          format: double
          type: number
      title: LocationDTO
      type: object
...
    ReservationDTOReq:
      example:
        reservationTime:
          hours: 2
          minutes: 9
          id: 7
        client:
          fname: fname
          lname: lname
          address:
            zip: zip
            city: city
            street: street
            id: 0
            state: state
            apartement: apartement
          deleted: true
          role: role
          phone: phone
          created: 2000-01-23
          dob: 2000-01-23
          imageURL: imageURL
          id: 0
          updated: 2000-01-23
          email: email
        location:
          latitude: 5.962133916683182
          id: 1
          longitude: 5.637376656633329
        barber:
          fname: fname
          lname: lname
          address:
            zip: zip
            city: city
            street: street
            id: 0
            state: state
            apartement: apartement
          deleted: true
          role: role
          phone: phone
          created: 2000-01-23
          dob: 2000-01-23
          imageURL: imageURL
          id: 0
          updated: 2000-01-23
          email: email
        haircut:
          estimatedTime: estimatedTime
          price: 0
          imageURL: imageURL
          description: description
          id: id
          title: title
        id: 6
        reservationDate: 2000-01-23
        status: status
      properties:
        barber:
          $ref: '#/components/schemas/UserDTOReq'
        client:
          $ref: '#/components/schemas/UserDTOReq'
        haircut:
          $ref: '#/components/schemas/HaircutDTO'
        id:
          format: int64
          type: integer
        location:
          $ref: '#/components/schemas/LocationDTO'
        reservationDate:
          format: date
          type: string
        reservationTime:
          $ref: '#/components/schemas/ReservationTimeDTO'
        status:
          type: string
      title: ReservationDTOReq
      type: object
    ReservationDTORes:
      example:
        reservationTime:
          hours: 2
          minutes: 9
          id: 7
        client:
          fname: fname
          address:
            zip: zip
            city: city
            street: street
            id: 0
            state: state
            apartement: apartement
          role: role
          created: 2000-01-23
          fullName: fullName
          lname: lname
          deleted: true
          phone: phone
          dob: 2000-01-23
          imageURL: imageURL
          id: 0
          updated: 2000-01-23
          email: email
        location:
          latitude: 5.962133916683182
          id: 1
          longitude: 5.637376656633329
        barber:
          fname: fname
          address:
            zip: zip
            city: city
            street: street
            id: 0
            state: state
            apartement: apartement
          role: role
          created: 2000-01-23
          fullName: fullName
          lname: lname
          deleted: true
          phone: phone
          dob: 2000-01-23
          imageURL: imageURL
          id: 0
          updated: 2000-01-23
          email: email
        haircut:
          estimatedTime: estimatedTime
          price: 0
          imageURL: imageURL
          description: description
          id: id
          title: title
        id: 6
        reservationDate: 2000-01-23
        status: status
      properties:
        barber:
          $ref: '#/components/schemas/UserDTORes'
        client:
          $ref: '#/components/schemas/UserDTORes'
        haircut:
          $ref: '#/components/schemas/HaircutDTO'
        id:
          format: int64
          type: integer
        location:
          $ref: '#/components/schemas/LocationDTO'
        reservationDate:
          format: date
          type: string
        reservationTime:
          $ref: '#/components/schemas/ReservationTimeDTO'
        status:
          type: string
      title: ReservationDTORes
      type: object
...
    ReservationTimeDTO:
      example:
        hours: 2
        minutes: 9
        id: 7
      properties:
        hours:
          format: int32
          type: integer
        id:
          format: int64
          type: integer
        minutes:
          format: int32
          type: integer
      title: ReservationTimeDTO
      type: object
...
    UserDTOReq:
      example:
        fname: fname
        lname: lname
        address:
          zip: zip
          city: city
          street: street
          id: 0
          state: state
          apartement: apartement
        deleted: true
        role: role
        phone: phone
        created: 2000-01-23
        dob: 2000-01-23
        imageURL: imageURL
        id: 0
        updated: 2000-01-23
        email: email
      properties:
        address:
          $ref: '#/components/schemas/AddressDTO'
        created:
          format: date
          type: string
        deleted:
          type: boolean
        dob:
          format: date
          type: string
        email:
          type: string
        fname:
          type: string
        id:
          format: int64
          type: integer
        imageURL:
          type: string
        lname:
          type: string
        phone:
          type: string
        role:
          type: string
        updated:
          format: date
          type: string
      title: UserDTOReq
      type: object
    UserDTORes:
      example:
        fname: fname
        address:
          zip: zip
          city: city
          street: street
          id: 0
          state: state
          apartement: apartement
        role: role
        created: 2000-01-23
        fullName: fullName
        lname: lname
        deleted: true
        phone: phone
        dob: 2000-01-23
        imageURL: imageURL
        id: 0
        updated: 2000-01-23
        email: email
      properties:
        address:
          $ref: '#/components/schemas/AddressDTO'
        created:
          format: date
          type: string
        deleted:
          type: boolean
        dob:
          format: date
          type: string
        email:
          type: string
        fname:
          type: string
        fullName:
          type: string
        id:
          format: int64
          type: integer
        imageURL:
          type: string
        lname:
          type: string
        phone:
          type: string
        role:
          type: string
        updated:
          format: date
          type: string
      title: UserDTORes
      type: object
...
  securitySchemes:
    JWT:
      in: header
      name: JWT
      type: apiKey