# src/main/java/ca/ghostteam/springulart/dto/LocationDTO.java
package ca.ghostteam.springulart.dto;

import lombok.Data;

/**
 * @author     
 * @version 1.0
 * @since 2022-03-29
 */
@Data
public class LocationDTO {
    private long id;
    private Double latitude;
    private Double longitude;
}